moved viewmodels to presenter and succesfully compiled
