+++ Warkey.Core/Presenter/AutoChatViewModel.cs
@@ -4,9 +4,8 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Input;
-using Warkey.Core.ViewModels;
 
-namespace Warkey.Core.ViewModels
+namespace Warkey.Core.Presenter
 {
     public class AutoChatViewModel : BaseViewModel
     {
