+++ Warkey.Core/Presenter/MainViewModel.cs
@@ -7,7 +7,7 @@ using System.Runtime.CompilerServices;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Warkey.Core.ViewModels
+namespace Warkey.Core.Presenter
 {
     public class MainViewModel : BaseViewModel
     {
