+++ Warkey.Core/Settings.cs
@@ -7,7 +7,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Warkey.Core.Services;
-using Warkey.Core.ViewModels;
+using Warkey.Core.Presenter;
 using Warkey.Infrastructure;
 
 namespace Warkey.Core
