+++ Warkey.Core/Warkey.Core.csproj
@@ -50,13 +50,13 @@
   <ItemGroup>
     <Compile Include="Commands\GameSettingsCommand.cs" />
     <Compile Include="Queries\GameSettingsQuery.cs" />
-    <Compile Include="Presenter\ViewModels\AutoChatsViewModel.cs" />
-    <Compile Include="Presenter\ViewModels\AutoChatViewModel.cs" />
-    <Compile Include="Presenter\ViewModels\BaseViewModel.cs" />
-    <Compile Include="Presenter\ViewModels\LoadGameViewModel.cs" />
-    <Compile Include="Presenter\ViewModels\MainViewModel.cs" />
-    <Compile Include="Presenter\ViewModels\SettingsViewModel.cs" />
-    <Compile Include="Presenter\ViewModels\WarkeyViewModel.cs" />
+    <Compile Include="Presenter\AutoChatsViewModel.cs" />
+    <Compile Include="Presenter\AutoChatViewModel.cs" />
+    <Compile Include="Presenter\BaseViewModel.cs" />
+    <Compile Include="Presenter\LoadGameViewModel.cs" />
+    <Compile Include="Presenter\MainViewModel.cs" />
+    <Compile Include="Presenter\SettingsViewModel.cs" />
+    <Compile Include="Presenter\WarkeyViewModel.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="MainServices.cs" />
     <Compile Include="Services\UpdateService.cs" />
